+++ Sources/Sentry/SentryTracer.m
@@ -123,7 +123,7 @@ static BOOL appStartMeasurementRead;
 
     SentrySpan *child = [[SentrySpan alloc] initWithTransaction:self context:context];
     @synchronized(self.children) {
-        [self.children addObject:child];
+        [_children addObject:child];
     }
 
     return child;